requirements.txt
openai-whisper==2024.7.0
ffmpeg-python==0.2.0